LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-animations/animation-base-response-001-expected.txt


FAIL em units respond to font-size animation assert_equals: expected "15px" but got "1px"
FAIL ex units respond to font-size animation assert_equals: expected "6.703125px" but got "0.4375px"
FAIL var() references respond to custom property animation assert_equals: expected "20px" but got "0px"
